dev/mcp-claude/weather_service/pyproject.toml
[project]
name = "weather-service"
version = "0.1.0"
description = "Weather Service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "httpx>=0.28.1",
 "mcp>=1.1.0",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "Christopher Tavolazzi"
email = "ctavolazzi@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
weather-service = "weather_service:main"
